test(ui): split execution purpose regression coverage
